requirements.txt
requests
python-dotenv
google.genai
